nootropia-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
celery
redis
pyalex
python-dotenv
bcrypt
pydantic[email]